requirements.txt
scrapy==2.11.0
selenium==4.16.0
rich==13.7.0
asyncio==3.4.3
webdriver-manager==4.0.1
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3
urllib3==2.1.0
pytz==2023.3.post1
six==1.16.0
idna==3.6
charset-normalizer==3.3.2
websockets==12.0
protobuf==3.20.3
google-protobuf==3.20.3
python-dotenv==1.0.0